requirements.txt
streamlit
openai-whisper
sentence-transformers
supabase
psycopg2-binary
requests
python-dotenv
torch